Add section on updating from single-DES
There are, unfortunately, still some single-DES deployments out
there.  Try to help them along by documenting a procedure for
migrating to stronger crypto.

The texinfo install guide had a section on "upgrading", but it was
not really suitable for direct import into a RST document.  For one,
it gave a high profile to the on-disk incompatibilities in upgrades
to 1.1 and 1.2.  It also was driven at upgrading *to* triple-des (or RC4),
which are something of a dead-end.  This new text attempts to be more
general and applicable to today's environment.