De-conditionalize Camellia code
The Camellia enctypes and cksumtypes have received IANA assignments.
Add #defines using those assignments to krb5.h, remove the CAMELLIA
conditional, and enable testing code as appropriate.

The Camellia draft has not received an RFC number yet, so there is no
Doxygen markup for the enctype and cksumtype #defines.  That can be
added once the RFC number is known.