Disconnect the texinfo admin guide from the build
Its content has been migrated to or superseded by the reST documentation,
essentially entirely in krb_admins.

A few portions of the texinfo document are simply no longer relevant
and do not need to be migrated.  In particular:
Information about reporting bugs lives on k5wiki.kerberos.org.
General Kerberos concepts/introduction will be elsewhere in the tree.
We do not need to document the time zones accepted by kadmin.
We do not need a table of the various error codes and strings in our
formal documentation.
A complete description of the layout of our source tree is not useful
or relevant to most Kerberos administrators.

ticket: 7408