Make glue for building PDFs
sphinx-build's latex output engine creates a subdirectory with
various latex files, and a Makefile.  The generated Makefile assumes
gmake, which we do not.  The logic needed in this makefile is rather
simple, so we just include it in src/doc/Makefile.in, even if we
do need a rather complicated shell expression to work in the subdirectory.

(cherry picked from commit f06725b7a98c16630f80d8cc1dd6386a7d21a058)

ticket: 7433