Move Release tag into footer for Sphinx HTML
The current position of the Release tag is in the body of the main page
of the Sphinx html.  Move it to the footer.  This way the release number
can be viewed from every page.

(cherry picked from commit 3cd8934659c4558a5144191716085d094b08d44e)

ticket: 7462
version_fixed: 1.11
status: resolved