Make sphinx warnings fatal for doc build
We currently do not have any warnings.  Let us keep it that way by
making warnings fatal in maintainer-mode (and configurable on the
buildslaves).  Using sphinx-build -W also causes errors to be reported
in the exit status and picked up by make, which is quite useful.

In order to allow the build bot to use -W but end-users to not use it,
SPHINX_ARGS must be passed on the command line; it cannot be set by
the convenience target 'htmlsrc'.  Document this.

(cherry picked from commit 7b6758a09f92ccc386022bcc86e3a5fea472fdce)

ticket: 7468
version_fixed: 1.11
status: resolved