Fix spin-loop bug in k5_sendto_kdc
In the second part of the first pass over the server list, we passed
the wrong list pointer to service_fds, causing it to see only a subset
of the server entries corresponding to sel_state.  This could cause
service_fds to spin if an event is reported on an fd not in the
subset.

(cherry picked from commit 2b06a22f7fd8ec01fb27a7335125290b8ceb6f18)

ticket: 7454
version_fixed: 1.10.4
status: resolved