Access keys for the KfW ribbon interface
Improve accessibility by actually enabling access keys for ribbon
elements (tap alt and follow the onscreen hints for keys to press),
instead of just underlining a letter in the name of each element.

Supply an underlined letter in the text of each element, corresponding
to this access key, even if there is not a shortcut key bound to that
element.  While here, fix conflicting assignment to 'R' on the 'options'
tab (between "Renewable Until" and "Automatic Ticket Renewal") by
making "Automatic Ticket Renewal" use 'T'.  Microsoft's UI recommendations
seem to say that access keys should be easy to locate when searching
through the menu, and thus using the first letter of the first or
second word is advisable.

The Ribbon XML Reference seems to indicate that these elements should
be "keytip" elements, but MSVS creates "keys" elements, which seem
to work, whereas "keytip" does not.  Apparently 'F' is standard for
the application button menu (which contains exit).  Access keys work
somewhat poorly for us in this menu, as they appear on top of the text
of the menu items, since we have no icons here.

ticket: 7441
tags: pullup
target_version: 1.10.4