Make resources.rst more useful to non-devs
Reorder the IRC channel listing so #kerberos is first.  (Developers
form a smaller part of our audience for this documentation set.)
Remove some details that are available on the wiki and not of interest
to non-developers.

Add a pointer to the wiki.

Fix minor grammar errors in Archives.

ticket: 7482 (new)
target_version: 1.11
tags: pullup