Make building docs easier in an unconfigured tree
Instead of requiring the user to specify srcdir, top_srcdir, PYTHON,
and possibly more in the future, make an 'htmlsrc' target that does so
for them.

This also lets us do the clean in the same step, so that only one
command is required of the user.

ticket: 7491 (new)
tags: pullup
target_version: 1.11