Use an empty challenge for the password question
If a question's challenge is NULL, it is unnecessarily difficult for a
responder callback to detect whether it was asked.  So it's better to
use an empty challenge when there is no challenge data to communicate.
Do this for the "password" question.

(cherry picked from commit 70f2d9a093c71624269b2317c62ad0993126bc40)

ticket: 7499
version_fixed: 1.11
status: resolved