Add copyright footer to HTML docs
The technique we use for inserting the feedback link in the footer
overrides the Sphinx basic/layout.html and agogo/layout.html footers
in a way that prevents us from getting the copyright link footer.
Copy the relevant part of the Sphinx basic/layout.html for now.

Add a copyright.rst that links to mitK5license.rst.

ticket: 7510 (new)
target_version: 1.11
tags: pullup