Don't output trailing whitespace in depfix.pl
All of our deps files have a trailing space in the first line,
originating from depfix.pl.  This space triggers the whitespace
checker when generating a new deps file.  Get rid of the trailing
whitespace in the script; the change will be reflected in the deps
files the next time they are regenerated.