Allow multi-hop SAM-2 exchanges
Prior to 1.11, it was possible to do SAM-2 preauth exchanges with
multiple hops by sending repeated preauth-required errors with
different challenges (which is not the way multi-hop exchanges are
described in RFC 6113, but it can still work).  This stopped working
when SAM-2 was converted to a built-in module because of the use_count
field.  Disable the use count for SAM-2 specifically.

ticket: 7571
version_fixed: 1.11.1
status: resolved